AI/Scene_Segmentation/requirements_last_scene.txt
torch>=1.12.0
torchvision>=0.13.0
Pillow>=8.0.0
tqdm>=4.60.0
scikit-learn>=1.0.0
numpy>=1.21.0 